// src/app/modules/gerente/pages/logs-sistema/models/TipoLogEnum.ts
export enum TipoLogEnum {
  LOGIN = 'LOGIN',
  REGISTRO = 'REGISTRO',
  ERRO = 'ERRO',
  MOVIMENTACAO_ENTRADA = 'MOVIMENTACAO ENTRADA',
  MOVIMENTACAO_SAIDA = 'MOVIMENTACAO SAIDA',
  SE_DESATIVOU = 'DESATIVOU SEU PERFIL',
  CANCELOU_CONSULTA = 'CANCELOU A CONSULTA',
  SOLICITOU_CONSULTA = 'SOLICITOU A CONSULTA',
  DEFERIU_CONSULTA = 'DEFERIU A CONSULTA',
  CONSULTA_INICIADA = 'CONSULTA INICIADA',
  CONSULTA_FINALIZADA = 'CONSULTA FINALIZADA',
  DESATIVOU_PERFIL = 'DESATIVOU PERFIL',
  INDEFERIU_CONSULTA = 'INDEFERIU A CONSULTA',
}